refactorizando nombres de variables y modilizando los componentes
