+++ src/Components/Loaders/styles.scss
@@ -19,5 +19,8 @@ $heightHeader: 60px;
   &-message {
     font-size: 25px;
     margin-bottom: 20px;
+    background: white;
+    padding: 10px;
+    border-radius: 10px;
   }
 }
