+++ src/Components/index.js
@@ -1,0 +1,13 @@
+import Breadcrumb from './Breadcrumb/breadcrumb.component'
+import Button from './Button/button'
+import Card from './Cards/card'
+import Error from './Error/error.component'
+import Loading from './Loaders/loading.component'
+
+export {
+  Breadcrumb,
+  Button,
+  Card,
+  Error,
+  Loading
+}
