+++ src/modules/SearchBar/search.component.js
@@ -1,11 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import './search.styles.scss'
-import Button from '../../Components/Button/button'
 import { Link } from 'react-router-dom'
-
-const MeliLogo = 'https://res.cloudinary.com/geraldeveloper27/image/upload/v1623710922/assets/Logo_MLx2.png'
-const searchIcon = 'https://res.cloudinary.com/geraldeveloper27/image/upload/v1623710922/assets/ic_Search.png'
+import { Button } from '../../Components'
+import { searchIcon, MeliLogo } from '../../utils'
+import './search.styles.scss'
 
 const SearchForm = ({ searchCallback, onChangeCallback, searchInput }) => (
   <form onSubmit={searchCallback}>
@@ -18,11 +16,11 @@ const SearchForm = ({ searchCallback, onChangeCallback, searchInput }) => (
 
     <Button type="submit" styleType="search" title={
       <img src={searchIcon} alt="search-icon" />
-    }/>
+    } />
   </form>
 )
 
-const SearchBarView = ({searchCallback, onChangeCallback, searchInput}) => (
+const SearchBarView = ({ searchCallback, onChangeCallback, searchInput }) => (
   <header className="nav-header">
     <div className="nav-header-container">
       <Link to="/">
@@ -53,6 +51,5 @@ SearchBarView.defaultProps = {
   searchInput: '',
 
 }
-
 
 export default SearchBarView
