+++ src/utils/images.js
@@ -1,0 +1,8 @@
+
+
+const shippingIcon = 'https://res.cloudinary.com/geraldeveloper27/image/upload/v1623710922/assets/ic_shipping2x.png'
+const MeliLogo = 'https://res.cloudinary.com/geraldeveloper27/image/upload/v1623710922/assets/Logo_MLx2.png'
+const searchIcon = 'https://res.cloudinary.com/geraldeveloper27/image/upload/v1623710922/assets/ic_Search.png'
+
+
+export { shippingIcon, MeliLogo, searchIcon }
